Матеріали/32_Flask_Парсинг даних. JSON/2_парсинг/pyproject.toml
[tool.poetry]
name = "32-flask-парсинг даних- json"
version = "0.1.0"
description = ""
authors = ["Oleksandr Kondratiuk <alex_kondr@outlook.com>"]
readme = "README.md"
packages = [{include = "32_flask_парсинг даних"}]

[tool.poetry.dependencies]
python = "^3.10"
flask-wtf = "^1.2.1"
sqlalchemy = "^2.0.35"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
